Refactored to use tags for playwright
