+++ frontend/fingertips-frontend/package.json
@@ -23,10 +23,10 @@
     "test-ci": "jest --ci --silent",
     "open-jest-coverage": "open coverage/lcov-report/index.html",
     "test-ui-local-mocks": "playwright test isolated_ui_tests --ignore-snapshots",
-    "test-e2e-local-docker": "cross-env MOCK_SERVER=false npm run start:all && playwright test fully_integrated_e2e_tests --ignore-snapshots || true && npm run stop:all",
-    "test-e2e-ci": "MOCK_SERVER=false playwright test fully_integrated_e2e_tests",
-    "test-e2e-cd": "MOCK_SERVER=false playwright test fully_integrated_e2e_tests ai_search_tests",
-    "test-e2e-ui-mode": "MOCK_SERVER=false playwright test fully_integrated_e2e_tests --ui --ignore-snapshots",
+    "test-e2e-local-docker": "cross-env MOCK_SERVER=false npm run start:all && playwright test fully_integrated_e2e_tests --grep @ci --ignore-snapshots || true && npm run stop:all",
+    "test-e2e-ci": "MOCK_SERVER=false playwright test fully_integrated_e2e_tests --grep @ci",
+    "test-e2e-cd": "MOCK_SERVER=false playwright test fully_integrated_e2e_tests ai_search_tests --grep \"@ci|@cd\"",
+    "test-e2e-ui-mode": "MOCK_SERVER=false playwright test fully_integrated_e2e_tests --grep @ci --ui --ignore-snapshots",
     "prettier": "prettier --write .",
     "prettier-ci": "prettier . --check"
   },
